requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.0.4
django-bootstrap4==24.3
django-crispy-forms==2.1
django-phonenumber-field==7.3.0
phonenumbers==8.13.36
soupsieve==2.5
sqlparse==0.5.0
